update nuget packages; move correlationcontext and hosting classes back to common and reference from there
