+++ Repository.psm1
@@ -1,0 +1,91 @@
+# these values are referenced in scripts that use dotnet ef commands
+Function Get-RepositoryConfiguration {
+	$configfile = "repository.json"
+	if (!(Test-Path $configfile)) {
+		throw "config file $configfile not found"
+	}
+	$config = Get-Content $configfile -raw | ConvertFrom-Json 
+
+	$configfile = "repository.local.json"
+	if (Test-Path $configfile) {
+		$local = Get-Content $configfile -raw | ConvertFrom-Json 
+		Merge-Json $config $local
+	}
+	
+	#TODO: hack for now to get secrets into $config
+	if ((Test-Path 'env:SONAR_TOKEN')) {
+		$config.sonar.token = $env:SONAR_TOKEN
+	}
+
+	$envconfig = [System.Environment]::GetEnvironmentVariables() | convertto-json | convertfrom-json
+	Merge-Json $config $envconfig
+
+	return $config
+}
+
+Function Merge-Json ($target, $source) {
+    $source.psobject.Properties | % {
+        if ($_.TypeNameOfValue -eq 'System.Management.Automation.PSCustomObject' -and $target."$($_.Name)" ) {
+            merge $target."$($_.Name)" $_.Value
+        } else {
+            $target | Add-Member -MemberType $_.MemberType -Name $_.Name -Value $_.Value -Force
+        }
+    }
+}
+
+function ConvertTo-Hashtable {
+    [CmdletBinding()]
+    [OutputType('hashtable')]
+    param (
+        [Parameter(ValueFromPipeline)]
+        $InputObject
+    )
+    process {
+        ## Return null if the input is null. This can happen when calling the function
+        ## recursively and a property is null
+        if ($null -eq $InputObject) {
+            return $null
+        }
+        ## Check if the input is an array or collection. If so, we also need to convert
+        ## those types into hash tables as well. This function will convert all child
+        ## objects into hash tables (if applicable)
+        if ($InputObject -is [System.Collections.IEnumerable] -and $InputObject -isnot [string]) {
+            $collection = @(
+                foreach ($object in $InputObject) {
+                    ConvertTo-Hashtable -InputObject $object
+                }
+            )
+            ## Return the array but don't enumerate it because the object may be pretty complex
+            Write-Output -NoEnumerate $collection
+        } elseif ($InputObject -is [psobject]) { ## If the object has properties that need enumeration
+            ## Convert it to its own hash table and return it
+            $hash = @{}
+            foreach ($property in $InputObject.PSObject.Properties) {
+                $hash[$property.Name] = ConvertTo-Hashtable -InputObject $property.Value
+            }
+            $hash
+        } else {
+            ## If the object isn't an array, collection, or other object, it's already a hash table
+            ## So just return it.
+            $InputObject
+        }
+    }
+}
+
+# Formats JSON in a nicer format than the built-in ConvertTo-Json does.
+function Format-Json([Parameter(Mandatory, ValueFromPipeline)][String] $json) {
+  $indent = 0;
+  ($json -Split '\n' |
+    % {
+      if ($_ -match '[\}\]]') {
+        # This line contains  ] or }, decrement the indentation level
+        $indent--
+      }
+      $line = (' ' * $indent * 2) + $_.TrimStart().Replace(':  ', ': ')
+      if ($_ -match '[\{\[]') {
+        # This line contains [ or {, increment the indentation level
+        $indent++
+      }
+      $line
+  }) -Join "`n"
+}
