+++ create-release.ps1
@@ -47,8 +47,8 @@ Function Get-Version {
 			}
 		}	
 	} else {
-		$a = Get-Content './repository.json' -raw | ConvertFrom-Json
-		return $a.version
+		$config = Get-Content './repository.json' -raw | ConvertFrom-Json
+		return $config.version
 	}
 }
 
@@ -74,19 +74,23 @@ Function Update-Version {
 		
 		git commit -m "update version" appveyor.yml
 	} else {
-		$a = Get-Content './repository.json' -raw | ConvertFrom-Json
-		$version = [version] $a.version
+		$config = Get-Content './repository.json' -raw | ConvertFrom-Json
+		$version = [version] $config.version
 		$versionStringIncremented =  [string] [version]::new(
 		  $version.Major,
 		  $version.Minor+1
 		)
-		$a.version = $versionStringIncremented
+		$config.version = $versionStringIncremented
 
-		$a | ConvertTo-Json -depth 32| set-content './repository.json'
+		$config | ConvertTo-Json -depth 32 | set-content './repository.json'
+		$config | ConvertTo-Json -depth 32 | Format-Json | Out-File -FilePath './repository.json' -Encoding utf8
 		
 		git commit -m "update version" ./repository.json
 	}
 }
+
+# common repository functions
+Import-Module .\Repository.psm1
 
 Invoke-Exe -cmd git -args "checkout master"
 Invoke-Exe -cmd git -args "pull"
@@ -109,7 +113,8 @@ if ($exists.Length -eq 0) {
 	git push --set-upstream origin $branch
 	
 	$remote = (git remote -v)
-	if ($remote -like "*github.com*") {
+	$ghexists = if (Get-Command "gh.exe" -ErrorAction SilentlyContinue) { $true } else { $false }
+	if ($remote -like "*github.com*" -and $ghexists) {
 		gh repo set-default
 		gh pr create --title "Release $version" --body "$releaseNotes" --base master
 	} else {
