+++ dependabot.ps1
@@ -2,7 +2,9 @@
 Param 
 (
     [Parameter(Mandatory = $false)][string]$package = "",
-	[Parameter(Mandatory = $false)][switch]$createpullrequest
+	[Parameter(Mandatory = $false)][string]$preupdateExpression = "",
+	[Parameter(Mandatory = $false)][switch]$createPullRequest,
+	[Parameter(Mandatory = $false)][switch]$ignoreChanges
 )
 
 $ErrorActionPreference = "Stop"
@@ -40,17 +42,41 @@ Function Invoke-Exe {
 }
 
 # check for uncommitted changed files
-git checkout -- *
-git status
+$changes = (git status --porcelain)
+if ($changes.Count -ne 0 -and -not $ignoreChanges.IsPresent) {
+	Write-Output "Exiting, sandbox has $($changes.Count) changes"
+	Write-Output $changes
+	exit 1
+}
 
 # make sure this is done from current develop branch
 Invoke-Exe git -args "checkout develop"
 Invoke-Exe git -args "pull"
 
+# check for branch first
+$bot = "BOT-{0:yyyyMMdd}" -f (Get-Date)
+$branch = "feature/$bot"
+if ($package -ne "") {
+	$branch = "feature/$bot-$package"
+}
+
+$exists = (git branch -r | sls $branch)
+if ($exists -ne $null) {
+	Write-Output "Exiting, $branch already exists"
+	exit 1
+}
+
 echo "prepping"
 
 .\clean.ps1 -quiet
 $result = check-result
+
+echo "preupdate"
+
+if ($preupdateExpression -ne "") {
+	echo "running $preupdateExpression"
+	Invoke-Expression "& $preupdateExpression"
+}
 
 echo "about to restore"
 
@@ -68,8 +94,8 @@ $result = check-result
 
 echo "checking to see if anything changed"
 
-$files = (git status *.csproj | grep "modified:" | wc -l)
-if ($files -ne "0") {
+$changes = (git status --porcelain)
+if ($changes.Count -ne 0) {
 	dotnet test src
 	$result = check-result
 
@@ -81,7 +107,7 @@ if ($files -ne "0") {
 		$branch = "feature/$bot-$package"
 	}
 
-	git add *.csproj
+	git add -u -u
 	git status
 	git checkout -b $branch
 	if ($package -eq "") { 
@@ -92,7 +118,8 @@ if ($files -ne "0") {
 	git push --set-upstream origin $branch
 
 	$remote = (git remote -v)
-	if ($remote -like "*github.com*") {
+	$ghexists = if (Get-Command "gh.exe" -ErrorAction SilentlyContinue) { $true } else { $false }
+	if ($remote -like "*github.com*" -and $ghexists) {
 		gh repo set-default
 		gh pr create --title "$bot" --body "$body" --base develop
 	} else {
@@ -102,8 +129,6 @@ if ($files -ne "0") {
 
 	.\clean.ps1
 	git checkout develop
-
-	
 } else {
 	echo "no files changed"
 }
