+++ src/Cortside.AspNetCore/Cortside.AspNetCore.csproj
@@ -13,20 +13,22 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="Cortside.Bowdlerizer" Version="6.0.59" />
-    <PackageReference Include="Cortside.Common.BootStrap" Version="6.2.406" />
-    <PackageReference Include="Cortside.Common.Cryptography" Version="6.2.406" />
-    <PackageReference Include="Cortside.Common.Json" Version="6.2.406" />
+    <PackageReference Include="Cortside.Common.BootStrap" Version="6.2.408" />
+    <PackageReference Include="Cortside.Common.Correlation" Version="6.3.443-release-6.3" />
+    <PackageReference Include="Cortside.Common.Cryptography" Version="6.2.408" />
+    <PackageReference Include="Cortside.Common.Hosting" Version="6.3.443-release-6.3" />
+    <PackageReference Include="Cortside.Common.Json" Version="6.2.408" />
     <PackageReference Include="Cortside.Common.Messages" Version="6.3.419-separate-aspnetcore" />
     <PackageReference Include="Cortside.Health" Version="6.0.86" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.26" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.32" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.1" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.1.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.1.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.8.14">
+    <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.10.48">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Serilog.AspNetCore" Version="8.0.0" />
+    <PackageReference Include="Serilog.AspNetCore" Version="8.0.2" />
     <PackageReference Include="Serilog.Bowdlerizer" Version="6.1.81" />
     <PackageReference Include="Serilog.Sinks.Seq" Version="6.0.0" />
   </ItemGroup>
