+++ update-scripts.ps1
@@ -1,0 +1,10 @@
+cp ..\coeus\shoppingcart-api\build.ps1
+cp ..\coeus\shoppingcart-api\clean.ps1
+cp ..\coeus\shoppingcart-api\create-release.ps1
+cp ..\coeus\shoppingcart-api\dependabot.ps1
+cp ..\coeus\shoppingcart-api\format.ps1
+cp ..\coeus\shoppingcart-api\generate-changelog.ps1
+cp ..\coeus\shoppingcart-api\update-nugetpackages.ps1
+cp ..\coeus\shoppingcart-api\update-targetframework.ps1
+cp ..\coeus\shoppingcart-api\src\.editorconfig .\src\.editorconfig
+cp ..\coeus\shoppingcart-api\Repository.psm1
